refactor: separeted DashboardAssetDetails components into single files
